+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.1.6
 
 - Use J/K to jump down/up 10 at a time in Newest Items
+- Hit b/B to open both permalink and comment thread
 
 # v0.1.5